Split dtype tests (nd/ndt).
